--- src/main/java/Main.java
@@ -1,12 +1,0 @@
-import controler.OrderController;
-
-public class Main {
-
-  public static void main(String[] args) {
-    System.out.println( "vending machine start." );
-
-    new OrderController().receive();
-
-    System.out.println( "vending machine finish." );
-  }
-}
