+++ src/main/java/api/Application.java
@@ -1,0 +1,12 @@
+package api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+  public static void main(String[] args) {
+    SpringApplication.run( Application.class, args );
+  }
+}
