+++ src/main/java/api/controler/form/PaymentForm.java
@@ -1,0 +1,12 @@
+package api.controler.form;
+
+import lombok.Data;
+
+import javax.validation.constraints.Positive;
+
+@Data
+public class PaymentForm {
+
+  @Positive
+  private Integer amount;
+}
