+++ src/main/java/api/controler/form/PurchaseForm.java
@@ -1,0 +1,12 @@
+package api.controler.form;
+
+import lombok.Data;
+
+import javax.validation.constraints.NotNull;
+
+@Data
+public class PurchaseForm {
+
+  @NotNull
+  private int drinkId;
+}
