+++ src/main/java/api/domain/entity/drink/DrinkDisplay.java
@@ -1,0 +1,15 @@
+package api.domain.entity.drink;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class DrinkDisplay {
+
+  private int drinkId;
+  private String drinkName;
+  private int drinkPrice;
+  private boolean isSaleable;
+  private String drinkTemperature;
+}
