+++ src/main/java/api/infrastructure/InMemory/CashStockRepositoryImpl.java
@@ -1,0 +1,32 @@
+package api.infrastructure.InMemory;
+
+import api.domain.entity.cash.Cash;
+import api.domain.entity.cash.CashStock;
+import api.repository.CashStockRepository;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+@Component
+public class CashStockRepositoryImpl implements CashStockRepository {
+
+  private static List<Cash> cashList = Arrays.asList(
+      new Cash( 10, 30 ),
+      new Cash( 50, 30 ),
+      new Cash( 100, 30 ),
+      new Cash( 500, 30 ),
+      new Cash( 1000, 30 )
+  );
+
+  @Override
+  public CashStock fetch() {
+    return new CashStock( new ArrayList<>( cashList ) );
+  }
+
+  @Override
+  public void store(CashStock cashStock) {
+    cashList = cashStock.getCashList();
+  }
+}
