+++ src/main/java/api/repository/CashStockRepository.java
@@ -1,0 +1,20 @@
+package api.repository;
+
+import api.domain.entity.cash.CashStock;
+
+public interface CashStockRepository {
+
+  /**
+   * 保有現金を取得.
+   *
+   * @return 保有現金
+   */
+  CashStock fetch();
+
+  /**
+   * 保有現金の更新.
+   *
+   * @param cashStock 保有現金
+   */
+  void store(CashStock cashStock);
+}
